requirements.txt
asgiref
channels[daphne]
Django
django-cors-headers
django-filter
djangorestframework
drf-spectacular
flake8
flake8-docstrings
tzlocal
